// src/infra/authorization-client/authorization-api-client/models/authorized-response.ts
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AuthorizedResponse
 */
export interface AuthorizedResponse {
    /**
     * 
     * @type {string}
     * @memberof AuthorizedResponse
     */
    'userId': string;
    /**
     * 
     * @type {boolean}
     * @memberof AuthorizedResponse
     */
    'isAuthorized': boolean;
}